Engine abstracted from hub
BelotRoom was the game driver, containing all the loops and logic to run the game, tightly coupled to the UI element of pushing updates to clients.

Engine extracted to its own class, containing various hooks that can be used to run some action, such as the pushing to client.

Engine takes in an IBelotObserver, such as LiveBelotObserver or SilentBelotObserver, which act on those hooks. LiveBelotObserver implements those hooks for pushing to client. No game state modifications in the observer.

The hub now only contains methods for directly receiving input from client and directly associated helpers. It looks up the associated BelotGameContext (game and observer) from the BelotGameRegistry, instantiates an engine (passing BelotGame and IBelotObserver), modifies whatever game state it needs to, and kicks off the engine loop.

OnConnected and OnDisconnected in hub refresh the IHubConnectedClients in the Observer. Registry is responsible for this refresh.

Disconnected tweaked to properly log and announce disconnections of spectators.

scoreTarget moved from BelotRoom to BelotGame